requirement.txt
aiohttp==3.9.3
aiosignal==1.3.1
appdirs==1.4.4
attrs==23.2.0
certifi==2024.2.2
colorama==0.4.6
dnspython==2.6.1
frozenlist==1.4.1
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
mysql-connector-python==8.3.0
nest-asyncio==1.6.0
numpy==1.26.4
pandas==2.2.0
prettytable==3.9.0
pyee==8.2.2
pymongo==4.7.3
pyodbc==5.1.0
pyppeteer==1.0.2
pyppeteer_stealth==2.7.4
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
SQLAlchemy==2.0.27
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.18
wcwidth==0.2.13
websockets==10.4
yarl==1.9.4
zipp==3.17.0